requirements.txt
aiohttp==3.8.1
aiohttp-apispec==2.2.3
aiomisc==16.0.13
aiosignal==1.2.0
alembic==1.8.0
apispec==3.3.2
async-timeout==4.0.2
asyncpg==0.25.0
asyncpgsa==0.27.1
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorlog==6.6.0
ConfigArgParse==1.5.3
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.16.0
multidict==6.0.2
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.37
urllib3==1.26.9
webargs==5.5.3
yarl==1.7.2
zipp==3.8.0
